+++ docs/release-notes.md
@@ -7,6 +7,7 @@
 
 * For mod authors:
   * Refactored how event handling works under the hood, particularly the new content API. This should have no effect on mod usage.
+  * In-game time changes are now logged in verbose mod.
   * Fixed error when loading a `.xnb` file through the old content API without the file extension.
   * Fixed asset propagation for player sprites not fully updating recolor masks in some cases.